+++ src/components/icons/index.tsx
@@ -70,3 +70,4 @@ export type { ProgressRingWithTrackProps } from './ProgressRingWithTrack';
 export { PromotedWidgetIcon } from './PromotedWidgetIcon';
 export { ServerIcon } from './ServerIcon';
 export { ClockIcon } from './ClockIcon';
+export { GearIcon } from './GearIcon';